Expose functionality as cell_area
