+++ tests/unit/test_utils.py
@@ -1,4 +1,5 @@
-from quadbin.utils import point_to_tile, point_to_tile_fraction
+import pytest
+from quadbin.utils import point_to_tile, point_to_tile_fraction, tile_area
 
 
 def test_point_to_tile_fraction():
@@ -19,3 +20,187 @@ def test_point_to_tile():
     assert point_to_tile(-175, -95, 2) == (0, 3, 2)
     assert point_to_tile(-175, 95, 2) == (0, 0, 2)
     assert point_to_tile(-175, 95, 2) == (0, 0, 2)
+
+
+@pytest.mark.parametrize(
+    "tile,expected",
+    [
+        ((0, 0, 0), 508164597540055.75),
+        ((1, 0, 1), 127516518279497.16),
+        ((0, 1, 1), 127516518279497.11),
+        ((0, 0, 2), 3354306636925.9023),
+        ((2, 1, 2), 60252354543011.56),
+        ((2, 2, 2), 60252354543011.56),
+        ((1, 3, 2), 3354306636925.902),
+        ((7, 0, 3), 405347921787.1127),
+        ((3, 1, 3), 1893932149859.4053),
+        ((7, 2, 3), 7940839890420.507),
+        ((1, 3, 3), 21531954963610.305),
+        ((2, 4, 3), 21531954963610.305),
+        ((6, 5, 3), 7940839890420.51),
+        ((7, 6, 3), 1893932149859.4014),
+        ((3, 7, 3), 405347921787.1137),
+        ((5, 0, 4), 68847216295.91176),
+        ((5, 1, 4), 150015920630.08636),
+        ((9, 2, 4), 324359672843.66473),
+        ((11, 3, 4), 689700245764.4768),
+        ((12, 4, 4), 1415535456168.1387),
+        ((3, 5, 4), 2703138561407.3833),
+        ((9, 6, 4), 4507012722233.37),
+        ((12, 7, 4), 6023040823252.668),
+        ((9, 8, 4), 6023040823252.664),
+        ((7, 9, 4), 4507012722233.373),
+        ((1, 10, 4), 2703138561407.385),
+        ((3, 11, 4), 1415535456168.1404),
+        ((1, 12, 4), 689700245764.4747),
+        ((2, 13, 4), 324359672843.66516),
+        ((8, 14, 4), 150015920630.0866),
+        ((4, 15, 4), 68847216295.91222),
+        ((2, 0, 5), 14160505713.702057),
+        ((18, 4, 5), 66957044671.40606),
+        ((4, 8, 5), 297109634630.5034),
+        ((14, 12, 5), 1008646311011.379),
+        ((14, 16, 5), 1549853128285.9084),
+        ((3, 20, 5), 780011091373.6311),
+        ((1, 24, 5), 207289317396.88107),
+        ((31, 28, 5), 45530712164.63003),
+        ((23, 0, 6), 3210498445.8824654),
+        ((34, 8, 6), 15204878147.827898),
+        ((18, 16, 6), 67965167366.12873),
+        ((17, 24, 6), 237461340707.57086),
+        ((52, 32, 6), 390274232638.2269),
+        ((32, 40, 6), 208783733329.56433),
+        ((4, 48, 6), 56765895703.36641),
+        ((27, 56, 6), 12538056116.820728),
+        ((81, 0, 7), 764328456.2867869),
+        ((4, 16, 7), 3622540506.9648323),
+        ((126, 32, 7), 16248126209.176235),
+        ((121, 48, 7), 57545410795.20297),
+        ((57, 64, 7), 97745045568.05563),
+        ((72, 80, 7), 53957641045.31359),
+        ((61, 96, 7), 14849099728.531345),
+        ((95, 112, 7), 3289546290.2144575),
+        ((138, 0, 8), 186467048.6535538),
+        ((19, 32, 8), 884078744.6901459),
+        ((193, 64, 8), 3971905493.817118),
+        ((201, 96, 8), 14160339910.8699),
+        ((27, 128, 8), 24447304143.76936),
+        ((148, 160, 8), 13711772787.766005),
+        ((148, 192, 8), 3797054754.548408),
+        ((174, 224, 8), 842465672.8105292),
+        ((14, 0, 9), 46050330.23509074),
+        ((196, 64, 9), 218372534.04884157),
+        ((137, 128, 9), 981880701.56478),
+        ((505, 192, 9), 3511936869.1768293),
+        ((491, 256, 9), 6112516398.900173),
+        ((213, 320, 9), 3455862981.7418623),
+        ((167, 384, 9), 960025220.6932801),
+        ((388, 448, 9), 213171250.2341762),
+        ((448, 0, 10), 11442422.668480542),
+        ((418, 128, 10), 54265162.87777075),
+        ((457, 256, 10), 244093752.50473356),
+        ((574, 384, 10), 874472509.6180239),
+        ((728, 512, 10), 1528172250.3897524),
+        ((675, 640, 10), 867463180.0306838),
+        ((994, 768, 10), 241361844.3184481),
+        ((289, 896, 10), 53615013.34380896),
+        ((2021, 0, 11), 2851875.6549530826),
+        ((1772, 256, 11), 13525475.896024764),
+        ((113, 512, 11), 60852040.82694955),
+        ((1877, 768, 11), 218179599.71270698),
+        ((1691, 1024, 11), 382045759.5605321),
+        ((869, 1280, 11), 217303430.58745173),
+        ((1443, 1536, 11), 60510553.14499288),
+        ((1262, 1792, 11), 13444207.546244394),
+        ((405, 0, 12), 711880.1518172809),
+        ((868, 512, 12), 3376278.427660119),
+        ((598, 1024, 12), 15191626.424900804),
+        ((2839, 1536, 12), 54490111.52470787),
+        ((1358, 2048, 12), 95511608.45103554),
+        ((268, 2560, 12), 54380590.292608134),
+        ((4, 3072, 12), 15148940.490946993),
+        ((46, 3584, 12), 3366119.894625659),
+        ((72, 0, 13), 177834.09803959142),
+        ((6103, 1024, 13), 843433.9941112136),
+        ((6838, 2048, 13), 3795236.185658999),
+        ((3122, 3072, 13), 13615681.06317344),
+        ((4012, 4096, 13), 23877912.64783286),
+        ((4822, 5120, 13), 13601990.906296989),
+        ((2784, 6144, 13), 3789900.44473846),
+        ((2225, 7168, 13), 842164.1778174711),
+        ((12132, 0, 14), 44441.5417196901),
+        ((11618, 2048, 14), 210779.09098281202),
+        ((3407, 4096, 14), 948475.4032619242),
+        ((1244, 6144, 14), 3403064.522131492),
+        ((8783, 8192, 14), 5969478.820397914),
+        ((1901, 10240, 14), 3401353.2524307016),
+        ((12185, 12288, 14), 947808.4356725625),
+        ((8108, 14336, 14), 210620.3639580674),
+        ((14689, 0, 15), 11108.263187119439),
+        ((1726, 4096, 15), 52684.84955532187),
+        ((5651, 8192, 15), 237077.15538302113),
+        ((19195, 12288, 15), 850659.1693733177),
+        ((9510, 16384, 15), 1492369.7462518667),
+        ((13382, 20480, 15), 850445.2606563419),
+        ((24101, 24576, 15), 236993.7844349456),
+        ((11169, 28672, 15), 52665.00867800594),
+        ((7292, 0, 16), 2776.8005542784613),
+        ((31685, 8192, 16), 13169.972161726404),
+        ((45857, 16384, 16), 59264.077539137535),
+        ((61186, 24576, 16), 212651.42262138167),
+        ((23968, 32768, 16), 373092.4391340615),
+        ((33538, 40960, 16), 212624.6840333054),
+        ((28645, 49152, 16), 59253.65617023244),
+        ((46470, 57344, 16), 13167.492052404968),
+        ((120813, 0, 17), 694.1669856280457),
+        ((74912, 16384, 17), 3292.33802293761),
+        ((11500, 32768, 17), 14815.368009842998),
+        ((30813, 49152, 17), 53161.18446727011),
+        ((17988, 65536, 17), 93273.10994449603),
+        ((116865, 81920, 17), 53157.84214334298),
+        ((74693, 98304, 17), 14814.065339102703),
+        ((130625, 114688, 17), 3292.0280091507925),
+        ((156343, 0, 18), 173.53760240416173),
+        ((16949, 32768, 18), 823.0651292155519),
+        ((207451, 65536, 18), 3703.7605831585784),
+        ((175018, 98304, 18), 13290.08721953591),
+        ((189809, 131072, 18), 23318.277496005037),
+        ((69596, 163840, 18), 13289.669429339378),
+        ((131755, 196608, 18), 3703.597749243366),
+        ((194095, 229376, 18), 823.0263775324452),
+        ((456398, 0, 19), 43.38388261963919),
+        ((255164, 65536, 19), 205.7638602834703),
+        ((496698, 131072, 19), 925.9299685042733),
+        ((254518, 196608, 19), 3322.4956929142404),
+        ((331731, 262144, 19), 5829.569374549895),
+        ((233238, 327680, 19), 3322.4434690965386),
+        ((415557, 393216, 19), 925.9096142673413),
+        ((346447, 458752, 19), 205.75901633678478),
+        ((991324, 0, 20), 10.845905913090053),
+        ((9697, 131072, 20), 51.440662297569574),
+        ((5314, 262144, 20), 231.48121997891653),
+        ((23851, 393216, 20), 830.6206592040279),
+        ((737269, 524288, 20), 1457.3923436601747),
+        ((572526, 655360, 20), 830.6141312184599),
+        ((52044, 786432, 20), 231.47867569299893),
+        ((58443, 917504, 20), 51.4400568296431),
+        ((1889412, 0, 21), 2.7114683806069366),
+        ((711837, 262144, 21), 12.860127728127763),
+        ((657038, 524288, 21), 57.87014596251734),
+        ((255406, 786432, 21), 207.6547568007845),
+        ((696547, 1048576, 21), 364.3480858946535),
+        ((400497, 1310720, 21), 207.6539408247401),
+        ((1523531, 1572864, 21), 57.86982793356335),
+        ((1211319, 1835008, 21), 12.860052049235128),
+        ((584591, 0, 22), 0.6778660821146193),
+        ((1353047, 524288, 22), 3.215027201573707),
+        ((2756290, 1048576, 22), 14.467516611460995),
+        ((2285964, 1572864, 22), 51.91363819170339),
+        ((3018232, 2097152, 22), 91.08702147901639),
+        ((637827, 2621440, 22), 51.91353620368485),
+        ((2343081, 3145728, 22), 14.4674768633863),
+        ((3745818, 3670016, 22), 3.2150177453530335),
+    ],
+)
+def test_tile_area(tile, expected):
+    assert tile_area(tile) == pytest.approx(expected, rel=1.5e-1)
